lab_09

+++ lab_09_3_1/src/sort.c
@@ -1,4 +1,3 @@
-#include <stdio.h>
 #include "../inc/sort.h"
 
 void swap(product *lhs, product *rhs)
@@ -8,10 +7,10 @@ void swap(product *lhs, product *rhs)
     *rhs = temp;
 }
 
-int sort_by_price(product *products, const int num_of_elements, const int given_price)
+int sort_by_price(product *products, const size_t num_of_elements, const unsigned int given_price)
 {
     int k = 0, count = 0;
-    for (int i = 0; i < num_of_elements; i++)
+    for (size_t i = 0; i < num_of_elements; i++)
     {
         if (products[i].price < given_price)
         {
@@ -19,6 +18,5 @@ int sort_by_price(product *products, const int num_of_elements, const int given_
             count++;
         }
     }
-    // printf("%d", count);
     return count;
 }
